Trying to end the job What was made - execution of the program was automatized (but not so beautiful as it could be) - sender now print the number of wanted characters put in the string - receivers identify itself,  realize where its substring starts and count the number of the occurrences of the wanted char on it. They also put this number on the appropiated space and show the number that is there (for reasons of debug)
Problems
- for precaution, the counters and the spaces in the memory for the
receiver should be long int at least, because of the superior limit of
the number of occurrences;
- when the code of sender (and of the receiver too, in the said debug
part) needs to sum the occurrences, the read number is negative, maybe
because its reading not just a space of a int, but a longer space (as
it can) and getting an overflowed number. I thought in a stream of bits
to resolve this.